move integerToHexColor to shared util file fiexes 1 of the previously mentioned errors
