+++ src/shared/util.ts
@@ -1,0 +1,3 @@
+export function integerToHexColor(integerColor: number) {
+    return '#' + integerColor.toString(16)
+}
